On branch OzDM's-First-Branch Changes to be committed: 	modified:   DCS_pcall.lua 	modified:   README.md
Added a bunch more commands from hoggit into the script. Stuff like inheritance is going to be an issue.
I think we should just replicate the functions if they are using a prefix specific prefix to run the command.
e.g. getByName() as it is returning the object itself, you cant just refer to the first variable as the objectID.

Stopped at StaticObjects